% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{corStruct}
\alias{corStruct}
\title{Simulation of Correlation Structure for Microbiome Data}
\usage{
corStruct(p, zero_prob = 0)
}
\arguments{
\item{p}{Integer, the number of variables to simulate.}

\item{zero_prob}{Numeric, in the range of \eqn{0} to \eqn{1}. The percentage
of zeroes in the non-diagonal lower-triangular matrix used to
generate the precision matrix.}
}
\value{
\code{corStruct} returns a list that contains
the following components:
\item{data.cor}{The correlation matrix.}
\item{data.precision.m}{The precision matrix, which is the
inverse of covariance matrix.}
\item{data.cov}{The covariance matrix.}
}
\description{
This function simulates the correlation between variables
within the microbiome data.
}
\examples{
data.cor.res = corStruct(p = 300, zero_prob = 0.7)

}
\references{
\insertRef{mcgregor2020mdine}{PLSDAbatch}
}
\author{
Yiwen Wang, Kim-Anh Lê Cao
}
