% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_batch.R
\name{deflate_mtx}
\alias{deflate_mtx}
\title{Matrix Deflation}
\usage{
deflate_mtx(X, t)
}
\arguments{
\item{X}{A numeric matrix to be deflated. It assumes that samples are
on the row, while variables are on the column. \code{NA}s are not allowed.}

\item{t}{A component to be deflated out from the matrix.}
}
\value{
A deflated matrix with the same dimension as the input matrix.
}
\description{
This function removes the variance of given component \code{t} from the
input matrix \code{X}. \deqn{\hat{X} = X - t (t^{\top}t)^{-1}t^{\top}X}
It is a built-in function of \code{PLSDA_batch}.
}
\examples{
# A built-in funciton of PLSDA_batch, not separately used.

}
\references{
\insertRef{barker2003partial}{PLSDAbatch}
}
\author{
Yiwen Wang, Kim-Anh Lê Cao
}
\keyword{Internal}
