% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_color.R
\name{pb_color}
\alias{pb_color}
\title{Color palette for PLSDA-batch}
\usage{
pb_color(num.vector)
}
\arguments{
\item{num.vector}{An integer vector specifying which color to use in the palette
(there are only 25 colors available).}
}
\value{
A vector of colors (25 colors max.)
}
\description{
The function outputs a vector of colors.
}
\examples{
my.colors = pb_color(1:5)

}
