% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simData_Gaussian}
\alias{simData_Gaussian}
\title{Data Simulation from Gaussian Distribution}
\usage{
simData_Gaussian(
  mean.batch,
  sd.batch = 1,
  mean.trt,
  sd.trt = 1,
  noise = 0.2,
  N,
  p_total,
  p_trt_relevant,
  p_bat_relevant,
  percentage_overlap_samples = 0.5,
  percentage_overlap_variables = 0.5,
  seeds = round(runif(1, 0, 999))
)
}
\arguments{
\item{mean.batch}{Numeric, the absolute mean of two batches respectively
in batch associated components.}

\item{sd.batch}{Numeric, the variability (standard deviation) of batch effects.}

\item{mean.trt}{Numeric, the absolute mean of two treatment groups
respectively in treatment associated components.}

\item{sd.trt}{Numeric, the variability (standard deviation) of
treatment effects.}

\item{noise}{Numeric, the variability (standard deviation) of
background noise.}

\item{N}{Integer, the number of samples.}

\item{p_total}{Integer, the number of variables.}

\item{p_trt_relevant}{Integer, the number of variables affected by
treatment effects.}

\item{p_bat_relevant}{Integer, the number of variables affected by
batch effects.}

\item{percentage_overlap_samples}{Numeric, in the range of \eqn{0}
to \eqn{1}, the percentage of samples in the first batch and first
treatment group.}

\item{percentage_overlap_variables}{Numeric, in the range of
\eqn{0} to \eqn{1}, the percentage of variables with both batch and
treatment effects given fixed number of variables with each effect.}

\item{seeds}{Integer. Default value is randomly selected from
\eqn{0} to \eqn{999}.}
}
\value{
\code{simData_Gaussian} returns a list that contains
the following components:
\item{data}{The simulated data.}
\item{cleanData}{The ground-truth data without the batch effect.}
\item{Y.trt}{The simulated outcome variable indicating the treatment
grouping information.}
\item{Y.bat}{The simulated outcome variable indicating the batch
grouping information.}
\item{true.trt}{The variables assigned with the treatment effect.}
\item{true.batch}{The variables assigned with the batch effect.}
\item{tComp_relevant}{The component simulated with the treatment effect.}
\item{bComp_relevant}{The component simulated with the batch effect.}
}
\description{
This function simulates data with treatment and batch effects
under different \code{batch x treatment designs} from Gaussian distribution
using component-based and multivariate strategy.
}
\examples{
## Not run:
## First example
## Balanced batch x treatment design
simulation1 <- simData_Gaussian(mean.batch = 7,
                                sd.batch = 8,
                                mean.trt = 3,
                                sd.trt = 2,
                                N = 40,
                                p_total = 300,
                                p_trt_relevant = 60,
                                p_bat_relevant = 150,
                                percentage_overlap_samples = 0.5,
                                percentage_overlap_variables = 0.5,
                                seeds = 1)

## Second example
## Unbalanced batch x treatment design
simulation2 <- simData_Gaussian(mean.batch = 7,
                                sd.batch = 8,
                                mean.trt = 3,
                                sd.trt = 2,
                                N = 40,
                                p_total = 300,
                                p_trt_relevant = 60,
                                p_bat_relevant = 150,
                                percentage_overlap_samples = 0.2,
                                percentage_overlap_variables = 0.5,
                                seeds = 1)
## End(Not run)
}
\references{
\insertRef{wang2020multivariate}{PLSDAbatch}
}
\author{
Yiwen Wang, Kim-Anh Lê Cao
}
