% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simData_mnegbinom}
\alias{simData_mnegbinom}
\title{Data Simulation from Multivariate Negative Binomial Distribution}
\usage{
simData_mnegbinom(
  batch.group = 2,
  mean.batch,
  sd.batch = 1,
  mean.trt,
  sd.trt = 1,
  mean.bg = 0,
  sd.bg = 0.5,
  N,
  p_total,
  p_trt_relevant,
  p_bat_relevant,
  percentage_overlap_samples = 0.5,
  percentage_overlap_variables = 0.5,
  data.cor,
  disp = 10,
  prob_zero = 0,
  seeds = round(runif(1, 0, 999))
)
}
\arguments{
\item{batch.group}{Integer, the number of batch groups. Only valid for 2 or 3
batch groups. Default value is 2.}

\item{mean.batch}{Numeric, the mean of batch effect coefficients for all
related variables.}

\item{sd.batch}{Numeric, the variability (standard deviation) of batch effect
coefficients for all related variables.}

\item{mean.trt}{Numeric, the mean of treatment effect coefficients for all
related variables}

\item{sd.trt}{Numeric, the variability (standard deviation) of treatment
effect coefficients for all related variables.}

\item{mean.bg}{Numeric, the mean of background noise.}

\item{sd.bg}{Numeric, the variability (standard deviation) of
background noise.}

\item{N}{Integer, the number of samples.}

\item{p_total}{Integer, the number of variables.}

\item{p_trt_relevant}{Integer, the number of variables affected by
treatment effects.}

\item{p_bat_relevant}{Integer, the number of variables affected by
batch effects.}

\item{percentage_overlap_samples}{Numeric, in the range of \eqn{0}
to \eqn{1}, the percentage of samples from one batch group in one
treatment group.}

\item{percentage_overlap_variables}{Numeric, in the range of \eqn{0}
to \eqn{1}, the percentage of variables with both batch and treatment
effects given fixed number of variables with each effect.}

\item{data.cor}{Correlation matrix. The correlation between each pair
of variables.}

\item{disp}{Numeric. The dispersion parameter of negative binomial
distribution.}

\item{prob_zero}{Numeric, in the range of \eqn{0} to \eqn{1}. The probability
of each value to be 0.}

\item{seeds}{Integer. Default value is randomly selected from
\eqn{0} to \eqn{999}.}
}
\value{
\code{simData_mnegbinom} returns a list that contains
the following components:
\item{data}{The simulated data.}
\item{cleanData}{The ground-truth data without the batch effect.}
\item{data_mu_matrix}{The \eqn{ln} value of the mean to model
negative binomial distribution for the simulated data.}
\item{clean_mu_matrix}{The \eqn{ln} value of the mean to model
negative binomial distribution for the ground-truth data.}
\item{Y.trt}{The simulated outcome variable indicating the treatment
grouping information.}
\item{Y.bat}{The simulated outcome variable indicating the batch
grouping information.}
\item{true.trt}{The variables assigned with the treatment effect.}
\item{true.batch}{The variables assigned with the batch effect.}
\item{zero_index}{The indexes of cells containing the resulting counts from
the Multivariate Negative Binomial Distribution that are inflated as zero.}
}
\description{
This function simulates data with treatment and batch effects
under different \code{batch x treatment designs} from multivariate negative
binomial distribution using quantile-quantile transformation between
multivariate normal and negative binomial distribution. This function can
generate two or three batch groups.
}
\examples{
## Not run:
## First example
## To simulate the correlation structure
data.cor.res = corStruct(p = 300, zero_prob = 0.7)

## Balanced batch x treatment design
simulation <- simData_mnegbinom(mean.batch = 7,
                                sd.batch = 8,
                                mean.trt = 3,
                                sd.trt = 2,
                                mean.bg = 0,
                                sd.bg = 0.2,
                                N = 40,
                                p_total = 300,
                                p_trt_relevant = 100,
                                p_bat_relevant = 200,
                                percentage_overlap_samples = 0.5,
                                percentage_overlap_variables = 0.5,
                                data.cor = data.cor.res$data.cor,
                                disp = 10,
                                prob_zero = 0,
                                seeds = 1)

## Second example
## Unbalanced batch x treatment design
simulation <- simData_mnegbinom(mean.batch = 7,
                                sd.batch = 8,
                                mean.trt = 3,
                                sd.trt = 2,
                                mean.bg = 0,
                                sd.bg = 0.2,
                                N = 40,
                                p_total = 300,
                                p_trt_relevant = 100,
                                p_bat_relevant = 200,
                                percentage_overlap_samples = 0.2,
                                percentage_overlap_variables = 0.5,
                                data.cor = data.cor.res$data.cor,
                                disp = 10,
                                prob_zero = 0,
                                seeds = 1)
## End(Not run)

}
\references{
\insertRef{wang2020multivariate}{PLSDAbatch}
}
\author{
Yiwen Wang, Kim-Anh Lê Cao
}
