% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_detection.R
\name{box_plot}
\alias{box_plot}
\title{Box Plot}
\usage{
box_plot(
  df,
  title = NULL,
  batch.legend.title = "Batch",
  ylab = "Value",
  x.angle = 0,
  x.hjust = 0.5,
  x.vjust = 0.5
)
}
\arguments{
\item{df}{A data frame used to draw the box plots.}

\item{title}{Character, the plot title.}

\item{batch.legend.title}{Character, the legend title of batches.}

\item{ylab}{Character, y-axis title.}

\item{x.angle}{Numeric, angle of x axis, in the range of \eqn{0} to \eqn{360}.}

\item{x.hjust}{Numeric, horizontal justification of x axis, in the range of \eqn{0} to \eqn{1}.}

\item{x.vjust}{Numeric, vertical justification of x axis, in the range of \eqn{0} to \eqn{1}.}
}
\value{
none
}
\description{
This function draws side-by-side box plots for each batch.
}
\examples{
data('AD_data')
ad.clr <- AD_data$EgData$X.clr
ad.batch = AD_data$EgData$Y.bat
ad.df <- data.frame(value = ad.clr[,1], batch = ad.batch)
box_plot(df = ad.df, title = 'OTU 12', x.angle = 30)

}
