% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_detection.R
\name{density_plot}
\alias{density_plot}
\title{Density Plot}
\usage{
density_plot(
  df,
  title = NULL,
  batch.legend.title = "Batch",
  xlab = "Value",
  title.hjust = 0.5
)
}
\arguments{
\item{df}{A data frame used to draw the density plots.}

\item{title}{Character, the plot title.}

\item{batch.legend.title}{Character, the legend title of batches.}

\item{xlab}{Character, x-axis title.}

\item{title.hjust}{Numeric, horizontal justification of the plot title, in the range of \eqn{0} to \eqn{1}.}
}
\value{
none
}
\description{
This function draws an overlap of multiple density plots for each batch.
}
\examples{
data('AD_data')
ad.clr <- AD_data$EgData$X.clr
ad.batch = AD_data$EgData$Y.bat
ad.df <- data.frame(value = ad.clr[,1], batch = ad.batch)
density_plot(df = ad.df, title = 'OTU 12')

}
