% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managing_batch_effects.R
\name{linear_regres}
\alias{linear_regres}
\title{Linear Regression}
\usage{
linear_regres(
  data,
  trt,
  batch = NULL,
  model = "linear model",
  p.adjust.method = "fdr"
)
}
\arguments{
\item{data}{A data frame that contains the response variables in the model.
Samples as rows and variables as columns.}

\item{trt}{A factor or a class vector for the treatment grouping information (categorical outcome variable).}

\item{batch}{A factor or a class vector for the batch grouping information (categorical outcome variable).}

\item{model}{The model to be used for fitting, either 'linear model' or 'linear mixed model'.}

\item{p.adjust.method}{The method to be used for p-value adjustment, either "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr" or "none".}
}
\value{
\code{linear_regres} returns a list that contains the following components:
\item{p_adjusted}{The adjusted p-values for each variable from the input data.}
\item{model}{The model used for fitting.}
\item{p.adjust.method}{The method used for p-value adjustment.}
}
\description{
This function fits linear regression on each variable of the input data including linear model,
linear mixed model and adjusts p-values for multiple comparisons using one of several methods.
}
\examples{
data('AD_data')
ad.clr <- AD_data$EgData$X.clr
ad.batch = AD_data$EgData$Y.bat
ad.trt = AD_data$EgData$Y.trt
ad.lm <- linear_regres(data = ad.clr, trt = ad.trt, batch = ad.batch, model = 'linear model')
ad.p.adj <- ad.lm$p_adjusted


}
