% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_detection.R
\name{Scatter_Density}
\alias{Scatter_Density}
\title{Principal Component Analysis (PCA) with Density Plots per Component}
\usage{
Scatter_Density(
  object,
  batch = NULL,
  trt = NULL,
  xlim = NULL,
  ylim = NULL,
  batch.legend.title = "Batch",
  trt.legend.title = "Treatment",
  density.lwd = 0.2,
  title = NULL,
  title.cex = 1.5,
  legend.cex = 0.7,
  legend.title.cex = 0.75
)
}
\arguments{
\item{object}{The object of class PCA.}

\item{batch}{A factor or a class vector for the batch grouping information (categorical outcome variable).}

\item{trt}{A factor or a class vector for the treatment grouping information (categorical outcome variable).}

\item{xlim}{A numeric vector of length 2, indicating the x coordinate ranges.}

\item{ylim}{A numeric vector of length 2, indicating the y coordinate ranges.}

\item{batch.legend.title}{Character, the legend title of batches.}

\item{trt.legend.title}{Character, the legend title of treatments.}

\item{density.lwd}{Numeric, the thickness of density lines.}

\item{title}{Character, the plot title.}

\item{title.cex}{Numeric, the size of plot title.}

\item{legend.cex}{Numeric, the size of legends.}

\item{legend.title.cex}{Numeric, the size of legend title.}
}
\value{
none
}
\description{
This function draws a PCA sample plot with density plots per principal component.
}
\examples{
data('AD_data')
ad.clr <- AD_data$EgData$X.clr
ad.pca.before <- pca(ad.clr, ncomp = 3, scale = TRUE)
ad.batch = AD_data$EgData$Y.bat
ad.trt = AD_data$EgData$Y.trt
Scatter_Density(project = ad.pca.before, batch = ad.batch, trt = ad.trt)


}
